protect objects in views to object owner

+++ backend/workouts/views.py
@@ -3,6 +3,7 @@ from datetime import date
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.contrib.auth.decorators import login_required
 from django.forms import formset_factory
+from django.http import HttpResponseForbidden
 from django.shortcuts import render, redirect, get_object_or_404
 from django.urls import reverse_lazy
 from django.views import generic
@@ -21,6 +22,11 @@ class WorkoutListView(LoginRequiredMixin, generic.ListView):
 
 class WorkoutDetailView(LoginRequiredMixin, generic.DetailView):
     model = models.Workout
+
+    def get_queryset(self):
+        qs = super().get_queryset()
+        qs = qs.filter(profile=self.request.user.profile)
+        return qs
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
@@ -43,6 +49,8 @@ class WorkoutCreateView(LoginRequiredMixin, generic.CreateView):
 @login_required
 def workout_delete(request, pk):
     workout = get_object_or_404(models.Workout, pk=pk)
+    if workout.user != request.user:
+        return HttpResponseForbidden("You are not allowed to delete this workout.")
     workout.delete()
     return redirect("workouts:index")
 
@@ -50,6 +58,9 @@ def workout_delete(request, pk):
 @login_required
 def workout_copy(request, pk):
     workout = get_object_or_404(models.Workout, pk=pk)
+    if workout.user != request.user:
+        # TODO maybe in the future, we can allow users to copy other users' workouts
+        return HttpResponseForbidden("You are not allowed to copy this workout.")
 
     # Shallow copy the workout
     new_workout = workout
@@ -83,6 +94,11 @@ def workout_copy(request, pk):
 class WorkoutExerciseDetailView(LoginRequiredMixin, generic.DetailView):
     model = models.WorkoutExercise
 
+    def get_queryset(self):
+        qs = super().get_queryset()
+        qs = qs.filter(workout__profile=self.request.user.profile)
+        return qs
+
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
         context["workout"] = context["workoutexercise"].workout
@@ -92,6 +108,11 @@ class WorkoutExerciseDetailView(LoginRequiredMixin, generic.DetailView):
 
 @login_required
 def workout_exercise_add(request, workout_id):
+    workout = get_object_or_404(models.Workout, pk=workout_id)
+    if workout.user != request.user:
+        return HttpResponseForbidden(
+            "You are not allowed to add an exercise to this workout."
+        )
     if request.method == "POST":
         form = forms.WorkoutExerciseForm(request.POST)
         if form.is_valid():
@@ -104,6 +125,8 @@ def workout_exercise_add(request, workout_id):
 @login_required
 def workout_exercise_delete(request, workout_exercise_id):
     workout_exercise = get_object_or_404(models.WorkoutExercise, pk=workout_exercise_id)
+    if workout_exercise.user != request.user:
+        return HttpResponseForbidden("You are not allowed to delete this exercise.")
     workout_exercise.delete()
     return redirect("workouts:detail", pk=workout_exercise.workout_id)
 
@@ -111,6 +134,10 @@ def workout_exercise_delete(request, workout_exercise_id):
 @login_required
 def set_add(request, workout_exercise_id):
     workout_exercise = get_object_or_404(models.WorkoutExercise, pk=workout_exercise_id)
+    if workout_exercise.user != request.user:
+        return HttpResponseForbidden(
+            "You are not allowed to add a set to this exercise."
+        )
     if request.method == "POST":
         form = forms.SetForm(request.POST)
         if form.is_valid():
@@ -122,28 +149,35 @@ def set_add(request, workout_exercise_id):
 
 @login_required
 def set_delete(request, pk):
-    set = get_object_or_404(models.Set, pk=pk)
-    set.delete()
-    return redirect("workouts:workout_exercise_detail", pk=set.workout_exercise_id)
+    set_ = get_object_or_404(models.Set, pk=pk)
+    if set_.user != request.user:
+        return HttpResponseForbidden(
+            "You are not allowed to add a set to this exercise."
+        )
+    set_.delete()
+    return redirect("workouts:workout_exercise_detail", pk=set_.workout_exercise_id)
 
 
 @login_required
 def set_copy(request, pk):
-    set = get_object_or_404(models.Set, pk=pk)
+    set_ = get_object_or_404(models.Set, pk=pk)
+    if set_.user != request.user:
+        return HttpResponseForbidden("You are not allowed to copy this set.")
     # A. Wei - By setting the pk to None, Django will create a new instance and autogen a new pk.
-    set.pk = None
-    set.save()
+    set_.pk = None
+    set_.save()
     return redirect("workouts:workout_exercise_detail", pk=set.workout_exercise_id)
-
-
-class SetDetailView(LoginRequiredMixin, generic.DetailView):
-    model = models.Set
 
 
 class SetUpdateView(LoginRequiredMixin, generic.UpdateView):
     model = models.Set
     fields = ["weight", "weight_unit", "repetitions", "seconds"]
     template_name_suffix = "_update_form"
+
+    def get_queryset(self):
+        qs = super().get_queryset()
+        qs = qs.filter(workout_exercise__workout__profile=self.request.user.profile)
+        return qs
 
     def get_success_url(self):
         return reverse_lazy(
